fix: automatically convert base_url from file url
The base url for prepend_url had undefined behavior if the url pointed
to a file (e.g., did not end with a /). Now this is detected and will be
converted.
`http://www.domain.com/path/to` -> `http://www.domain.com/path/to/`

This was necessary so the appended url would be added to the path.
`http://www.domain.com/path/to/` + `new_path` ->
`http://www.domain.com/path/to/new_path` instead of
`http://www.domain.com/path/tonew_path

https://github.com/alandtse/alexa_media_player/issues/2111